+++ sphinx/source/changelog.rst
@@ -3,8 +3,9 @@ Changelog
 
 5.1b1
 -----
-* Added the ``--ssub`` and ``--ofix`` options to :ref:`skool2bin.py` (for
-  parsing the skool file in ``@ssub`` or ``@ofix`` mode)
+* Added the ``--bfix``,  ``--ofix`` and ``--ssub`` options to
+  :ref:`skool2bin.py` (for parsing the skool file in ``@bfix``, ``@ofix`` and
+  ``@ssub`` mode)
 * The :ref:`R` macro now converts an anchor that matches the entry address to
   the format specified by the ``AddressAnchor`` parameter (making it easier to
   link to the first instruction in an entry when using a custom anchor format)